% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplot.R
\name{barFlip}
\alias{barFlip}
\title{To present a flip barplot}
\usage{
barFlip(data, x, y)
}
\arguments{
\item{data}{dataframe}

\item{x}{x}

\item{y}{y}
}
\value{
a flip barplot
}
\description{
To present a flip barplot
}
