% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPlot.R
\name{easyPlot}
\alias{easyPlot}
\title{Launch easyPlot interactive User Interface}
\usage{
easyPlot(data = "FALSE")
}
\arguments{
\item{data}{an optional data frame with at least two columns passed as a character string}
}
\description{
Launch easyPlot interactive User Interface
}
\details{
easyPlot()

data(iris)

easyPlot(data = "iris")
}
