% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPlot.R
\name{easyPlot}
\alias{easyPlot}
\title{Start an interactive User Interface easyPlot}
\usage{
easyPlot(data = "FALSE")
}
\arguments{
\item{data}{(character) an optional data frame with at least two columns passed as a character string}
}
\description{
Start an interactive User Interface easyPlot
}
\examples{
# easyPlot()

# easyPlot(data = "iris")
}
\author{
Michal Majka
}
