% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyPlot.R
\name{easyPlot}
\alias{easyPlot}
\title{Start an Interactive User Interface `easyPlot`}
\usage{
easyPlot(data = NULL)
}
\arguments{
\item{data}{(data.frame) an optional data frame with at least two columns passed as a character string}
}
\description{
Start an Interactive User Interface `easyPlot`
}
\examples{
# easyPlot()
# easyPlot(iris)
}
\author{
Michal Majka
}
