\name{reorder.cormat}
\alias{reorder.cormat}
\title{Re-ordered cluster matrix}
\usage{
timeline(data, x, y, color, title)
}
\arguments{
\item{matrix}{matrix}
}
\description{
\code{reorder.cormat} clusters correlational values of a given matrix.
}
\example{
m.mtcars = as.matrix(mtcars)
reorder.cormat(m.mtcars)

}
